requirements.txt
feedparser>=6.0.10
requests>=2.31.0
urllib3>=2.0.0
rich>=13.7.0
python-dateutil>=2.8.2
